Conflicts:
	Xihad.sln

+++ include/irrlicht/IParticleGravityAffector.h
@@ -19,13 +19,13 @@ public:
 
 	//! Set the time in milliseconds when the gravity force is totally lost
 	/** At that point the particle does not move any more. */
-	virtual void setTimeForceLost( f32 timeForceLost ) = 0;
+	virtual void setTimeForceLost( u32 timeForceLost ) = 0;
 
 	//! Set the direction and force of gravity in all 3 dimensions.
 	virtual void setGravity( const core::vector3df& gravity ) = 0;
 
 	//! Get the time in milliseconds when the gravity force is totally lost
-	virtual f32 getTimeForceLost() const = 0;
+	virtual u32 getTimeForceLost() const = 0;
 
 	//! Get the direction and force of gravity.
 	virtual const core::vector3df& getGravity() const = 0;
